% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{col_pop}
\alias{col_pop}
\title{Count the number of colonised patients}
\usage{
col_pop(admissions, colonisations, discharges, statuses, day, antibiotics)
}
\arguments{
\item{admissions}{a vector with days of admission of the patients.}

\item{discharges}{a vector with days of discharge of the patients.}

\item{statuses}{a vector the statuses on discharge of the patients}

\item{day}{the day for which the colonised number of colonised patients
is counted.}

\item{antibiotics}{a matrix where each row corresponds to a patient
and the entries correspond to a day when antibiotics were administered.}

\item{colonsiations}{a vector with days of colonisation of the patients.}
}
\description{
This function counts the number of colonised patients on and off antibiotics
on a given day.
}
