% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{count_false_negative_tests}
\alias{count_false_negative_tests}
\title{Count false negative tests}
\usage{
count_false_negative_tests(colonisations, statuses, test_results_negative,
  antibiotics)
}
\arguments{
\item{colonisations}{a vector with days of colonisation of the patients.}

\item{statuses}{a vector the statuses on discharge of the patients}

\item{test_results_negative}{a matrix where each row corresponds to a patient
and the entries correspond to a day with a negative test result.}

\item{antibiotics}{a matrix where each row corresponds to a patient
and the entries correspond to a day when antibiotics were administered.}
}
\description{
This function counts the number of false negative tests conducted with and
without antibioitics.
}
