% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{count_tests_forward_simulations}
\alias{count_tests_forward_simulations}
\title{Count positive and negative tests}
\usage{
count_tests_forward_simulations(rho_1, rho_2, colonisations, antibiotics,
  days_of_tests, n_patients)
}
\arguments{
\item{rho_1}{the test sensitivity in the absence of antibiotics.}

\item{rho_2}{the test sensitivity in the presence of antibiotics.}

\item{colonisations}{a vector of colonisation times of the patients.}

\item{antibiotics}{a matrix where each row corresponds to a patient
and the entries correspond to a day when antibiotics were administered.}

\item{days_of_tests}{a matrix where each row corresponds to a patient
and the entries correspond to a day when a test was conducted.}

\item{n_patients}{the total number of patients.}
}
\value{
This function returns a vector with two elements. The first number
is the total number of positive tests, the second number is the total
number of negative tests.
}
\description{
This function counts the positive and negative tests in the simulated
data generated by \code{run_forward_simulations()}.
}
