% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{count_true_positive_tests}
\alias{count_true_positive_tests}
\title{Count true positive tests.}
\usage{
count_true_positive_tests(test_results_positive, antibiotics)
}
\arguments{
\item{test_results_positive}{a matrix where each row corresponds to a patient
and the entries correspond to a day with a positive test result.}

\item{antibiotics}{a matrix where each row corresponds to a patient
and the entries correspond to a day when antibiotics were administered.}
}
\description{
This function counts the number of true positive tests conducted with and
without antibioitics. We assume perfect specificity, so each positive test
is a true positive test.
}
