% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{daily_col_prob_precalculated}
\alias{daily_col_prob_precalculated}
\title{Compute the probability of acquisition}
\usage{
daily_col_prob_precalculated(admissions, colonisations, discharges,
  statuses, antibiotics, cp_no_abx, cp_abx, beta, b, s, day, pt)
}
\arguments{
\item{admissions}{a vector with days of admission of the patients.}

\item{discharges}{a vector with days of discharge of the patients.}

\item{statuses}{a vector the statuses on discharge of the patients}

\item{antibiotics}{a matrix where each row corresponds to a patient
and the entries correspond to a day when antibiotics were administered.}

\item{cp_no_abx}{the number of colonised patients not on antibiotics
on the given day.}

\item{cp_abx}{the number of colonised patients on antibiotics
on the given day.}

\item{beta}{the transmission rate.}

\item{b}{the multiplicative effect of antibiotics on transmissibility.}

\item{s}{the multiplicative effect of antibiotics on susceptibility.}

\item{day}{the day for which the probability of acquisition is calculated.}

\item{pt}{the patient for which the probability of acquisition is calculated.}

\item{colonsiations}{a vector with days of colonisation of the patients.}
}
\description{
This function calculates the probability of acquisition for a given day
and patient using precalculated numbers of colonised patients on and off
antibiotics on the given day.
}
