% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_priors.R
\name{generate_priors}
\alias{generate_priors}
\title{Generate priors.}
\usage{
generate_priors(f_beta = NULL, f_b = NULL, f_s = NULL,
  f_rho_1 = NULL, f_rho_2 = NULL, p_beta = NULL, p_b = NULL,
  p_s = NULL, p_rho_1 = NULL, p_rho_2 = NULL)
}
\arguments{
\item{f_beta}{the function to be used as a prior of beta. The default
is Cauchy.}

\item{f_b}{the function to be used as a prior of b. The default
is Normal.}

\item{f_s}{the function to be used as a prior of s. The default
is Normal.}

\item{f_rho_1}{the function to be used as a prior of rho_1. The default
is Uniform.}

\item{f_rho_2}{the function to be used as a prior of rho_2. The default
is Uniform.}

\item{p_beta}{a vector with parameters for the prior of beta. The default
is c(0, 4).}

\item{p_b}{a vector with parameters for the prior of b. The default
is c(1, 0.5).}

\item{p_s}{a vector with parameters for the prior of s. The default
is c(1, 0.5).}

\item{p_rho_1}{a vector with parameters for the prior of rho_1. The default
is c(0, 1).}

\item{p_rho_2}{a vector with parameters for the prior of rho_2. The default
is c(0, 1).}
}
\description{
This function generates a list of functions and a list of parameters
to be used as priors in \code{run_mcmc()}. If nothing is supplied,
default priors are returned.
}
\examples{
## use default priors
generate_priors()

## modify prior for beta
generate_priors(f_beta = dnorm, p_beta = c(1, 0.5))
}
