% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{log_lik_importation}
\alias{log_lik_importation}
\title{Log likelihood of importation}
\usage{
log_lik_importation(nPatient, statuses, phi)
}
\arguments{
\item{nPatient}{the total number of patients.}

\item{statuses}{a vector the statuses on discharge of the patients}

\item{phi}{the probability of a patient beeing colonised on admission.}

\item{TP_no_abx}{the number of true positive tests without antibiotics.}
}
\description{
This function calculates the importation component of the overall log
likelihood.
}
