% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{log_lik_overall}
\alias{log_lik_overall}
\title{Overall log likelihood for one patient}
\usage{
log_lik_overall(admissions, discharges, colonisations, statuses, TP_no_abx,
  TP_abx, test_results_negative, antibiotics, beta, b, s, rho_1, rho_2,
  nPatient, phi, pt)
}
\arguments{
\item{admissions}{a vector with days of admission of the patients.}

\item{discharges}{a vector with days of discharge of the patients.}

\item{colonisations}{a vector with days of colonisation of the patients.}

\item{statuses}{a vector the statuses on discharge of the patients}

\item{TP_no_abx}{the number of true positive tests without antibiotics.}

\item{TP_abx}{the number of true positive tests with antibiotics.}

\item{test_results_negative}{a matrix where each row corresponds to a
patient and the entries correspond to a day with a negative test result.}

\item{antibiotics}{a matrix where each row corresponds to a patient
and the entries correspond to a day when antibiotics were administered.}

\item{beta}{the transmission rate.}

\item{b}{the multiplicative effect of antibiotics on transmissibility.}

\item{s}{the multiplicative effect of antibiotics on susceptibility.}

\item{rho_1}{the test sensitivity without antibiotics (baseline).}

\item{rho_2}{the test sensitivity with antibiotics.}

\item{nPatient}{the total number of patients.}

\item{phi}{the probability of a patient beeing colonised on admission.}

\item{pt}{the patient for which the overall log likelihood is computed.}

\item{beta}{the transmission rate.}
}
\description{
This function calculates the summand of the overall log likelihood that
changes when the status of one patient is updated. The constant summands
are ignored for efficiency.
}
