% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{log_lik_rho}
\alias{log_lik_rho}
\title{Log likelihood of test sensitivity}
\usage{
log_lik_rho(colonisations, statuses, TP_no_abx, TP_abx,
  test_results_negative, antibiotics, rho_1, rho_2)
}
\arguments{
\item{colonisations}{a vector with days of colonisation of the patients.}

\item{statuses}{a vector the statuses on discharge of the patients}

\item{TP_no_abx}{the number of true positive tests without antibiotics.}

\item{TP_abx}{the number of true positive tests with antibiotics.}

\item{test_results_negative}{a matrix where each row corresponds to a patient
and the entries correspond to a day with a negative test result.}

\item{antibiotics}{a matrix where each row corresponds to a patient
and the entries correspond to a day when antibiotics were administered.}

\item{rho_1}{the test sensitivity without antibiotics (baseline).}

\item{rho_2}{the test sensitivity with antibiotics.}
}
\description{
This function calculates the test sensitivity component of the overall log
likelihood.
}
