% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{log_lik_sources}
\alias{log_lik_sources}
\title{Log likelihood of sources}
\usage{
log_lik_sources(nPatients, nTypes, admission, colonisation, source,
  discharge, status, antibiotics, b)
}
\arguments{
\item{nPatients}{the total number of patients.}

\item{nTypes}{the number of differnt types of colonisations.}

\item{admission}{a vector with days of admission of the patients.}

\item{colonisation}{a vector with days of colonisation of the patients.}

\item{source}{a vector with the sources of the patients.}

\item{discharge}{a vector with days of discharge of the patients.}

\item{status}{a vector with the statuses on discharge of the patients.}

\item{antibiotics}{a matrix where each row corresponds to a patient
and the entries correspond to a day when antibiotics were administered.}

\item{b}{the multiplicative effect of antibiotics on transmissibility.}
}
\description{
This function calculates the source component of the overall log likelihood.
}
