% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{log_lik_transmission}
\alias{log_lik_transmission}
\title{Log likelihood of transmission}
\usage{
log_lik_transmission(admissions, colonisations, discharges, statuses,
  antibiotics, cp_no_abx, cp_abx, beta, b, s)
}
\arguments{
\item{admissions}{a vector with days of admission of the patients.}

\item{colonisations}{a vector with days of colonisation of the patients.}

\item{discharges}{a vector with days of discharge of the patients.}

\item{statuses}{a vector the statuses on discharge of the patients}

\item{antibiotics}{a matrix where each row corresponds to a patient
and the entries correspond to a day when antibiotics were administered.}

\item{cp_no_abx}{the number of colonised patients not on antibiotics
on the given day.}

\item{cp_abx}{the number of colonised patients on antibiotics
on the given day.}

\item{beta}{the transmission rate.}

\item{b}{the multiplicative effect of antibiotics on transmissibility.}

\item{s}{the multiplicative effect of antibiotics on susceptibility.}
}
\description{
This function calculates the transmission component of the overall log
likelihood.
}
