% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{onward_check}
\alias{onward_check}
\title{Get last possible day for colonisation}
\usage{
onward_check(pt, admissions, colonisations, discharges,
  first_positive_test, statuses, nPatients)
}
\arguments{
\item{pt}{the patient for which the onward transmission is calculated.}

\item{admissions}{a vector with days of admission of the patients.}

\item{colonisations}{a vector with days of colonisation of the patients.}

\item{discharges}{a vector with days of discharge of the patients.}

\item{first_positive_test}{a vector with the days of the first positive
test of the patients.}

\item{statuses}{a vector the statuses on discharge of the patients}

\item{nPatients}{the total number of patients.}
}
\description{
This function returns the last day on which a patient can become colonised.
If an onward transmission depends on that patient, the last possible day is
the day before the onward tranmission. If not, it is the minimum of the
first positive test and discharge.
}
