% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_dataset.R
\docType{data}
\name{patient_data_simulated}
\alias{patient_data_simulated}
\title{Simulated patient data.}
\format{a list containing 4 elements:
\describe{
  \item{patients}{a dataframe with four columns and 270 rows. Each row
  corresponds to a patient, the columns are admission, type,
  first_positive_test and discharge in days. If there is no type for a patient,
  type equals -1. If there is no positive test for a patient, first_positive_test
  equals 20000.}
  \item{test_results_positive}{a matrix with 270 rows and 30 columns. Each
  row corresponds to a patient. Entries correspond to the day of a positive
  test.}
   \item{test_results_negative}{a matrix with 270 rows and 30 columns. Each
  row corresponds to a patient. Entries correspond to the day of a negative
  test.}
   \item{antibiotics}{a matrix with 270 rows and 100 columns. Each
  row corresponds to a patient. Entries correspond to the day when an
  antibiotic was administered.}
}}
\usage{
patient_data_simulated
}
\description{
A dataset containing admission and discharge dates, test
results and antibiotic use of 270 patients over 90 days.
}
\keyword{datasets}
