% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mcmc.R
\name{run_mcmc}
\alias{run_mcmc}
\title{Run data augmented MCMC}
\usage{
run_mcmc(data = NULL, priors = NULL, configuration = NULL)
}
\arguments{
\item{data}{a list containing 4 elements:
\describe{
  \item{patients}{a dataframe with three columns. Each row
  corresponds to a patient, the columns are admission, first_positive_test
  and discharge. Entries are in days. If there is no positive test for a patient,
  first_positive_test equals 20000.}
  \item{test_results_positive}{a matrix. Each row corresponds to a patient.
  Entries correspond to the day of a positive test.}
   \item{test_results_negative}{a matrix. Each row corresponds to a patient.
   Entries correspond to the day of a negative test.}
   \item{antibiotics}{a matrix. Each row corresponds to a patient.
   Entries correspond to the day when an antibiotic was administered.}
}}

\item{priors}{a list containing two lists, functions and params, defining
the functions and parameters used as priors. If no input is supplied,
default priors are used. Default priors are described in the documentation
of \code{generate_priors()}. Custom priors can be generated using
\code{generate_priors()}.}

\item{configuration}{a list of settings. If no input is supplied, default
settings are used. Default settings are described in the documentation of
\code{generate_configuration()}. Custom settings can be generated using
\code{generate_configuration()}.}
}
\value{
The is a list of two dataframes. The first dataframe contains the
chains of the parameter values, with iterations stored as specified in the
configuratin. The second dataframe contains the chains of the statuses,
with iterations stored as specified in the configuration.
}
\description{
Run data augmented MCMC
}
\examples{
\dontrun{
#run example dataset with default priors and default settings.
run_mcmc()

#run example dataset with custom priors and custom settings.
my_priors <- generate_priors(f_beta = dnorm, p_beta = c(0.02,0.001))
my_configuration <- generate_configuration(n_iter = 300, n_burnin = 100)
run_mcmc(data = patient_data_simulated, priors = my_priors,
configuration = my_configuration)
}
}
