% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_onward_transmission}
\alias{update_onward_transmission}
\title{Get first onward transmission of a patient}
\usage{
update_onward_transmission(pt, source, colonisation, n_patients)
}
\arguments{
\item{pt}{an integer. the patient for which the first onward transmission
is returned.}

\item{source}{a vector. the sources of the patients.}

\item{colonisation}{a vector with days of colonisation of the patients.}

\item{n_patients}{the total number of patients.}
}
\description{
This function returns the first onward transmission for a given patient.
}
